conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.10.0=pypi_0
appdirs=1.4.4=pypi_0
argh=0.26.2=pyh9f0ad1d_1002
argon2-cffi=20.1.0=py37h8f50634_1
async_generator=1.10=py_0
attrs=20.2.0=pyh9f0ad1d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.1=py_0
blas=1.0=mkl
bleach=3.2.1=pyh9f0ad1d_0
brotlipy=0.7.0=py37h8f50634_1000
ca-certificates=2021.1.19=h06a4308_0
certifi=2020.6.20=pyhd3eb1b0_3
cffi=1.14.3=py37he30daa8_0
cfgv=3.2.0=pypi_0
chardet=3.0.4=py37hc8dfbb8_1007
click=7.1.2=pyh9f0ad1d_0
configparser=5.0.1=py_0
contextlib2=0.6.0.post1=pypi_0
cryptography=3.1.1=py37hb09aad4_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=py_2
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
distlib=0.3.1=pypi_0
docker-pycreds=0.4.0=py_0
entrypoints=0.3=py37hc8dfbb8_1001
expat=2.2.10=he6710b0_2
filelock=3.0.12=pypi_0
fontconfig=2.13.0=h9420a91_0
freetype=2.10.2=h5ab3b9f_0
gitdb=4.0.5=py_0
gitpython=3.1.9=py_0
glib=2.66.1=h92f7085_0
gql=0.1.0=py_0
graphql-core=3.1.2=py_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
icu=58.2=he6710b0_3
identify=1.5.5=pypi_0
idna=2.10=pyh9f0ad1d_0
importlib-metadata=2.0.0=py37hc8dfbb8_0
importlib_metadata=2.0.0=0
intel-openmp=2020.2=254
ipykernel=5.3.4=py37h43977f1_0
ipython=7.18.1=py37hc6149b9_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.5.1=py_1
jedi=0.17.2=py37hc8dfbb8_0
jinja2=2.11.2=pyh9f0ad1d_0
joblib=0.17.0=py_0
jpeg=9b=h024ee3a_2
json5=0.9.5=pyh9f0ad1d_0
jsonschema=3.2.0=py37hc8dfbb8_1
jupyter=1.0.0=py37_7
jupyter_client=6.1.7=py_0
jupyter_console=6.2.0=py_0
jupyter_core=4.6.3=py37hc8dfbb8_1
jupyterlab=2.2.9=py_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_server=1.2.0=py_0
kiwisolver=1.2.0=py37h99015e2_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.13.0.1=h8b12597_0
libsodium=1.0.18=h516909a_1
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
llvmlite=0.35.0=pypi_0
lz4-c=1.9.2=heb0550a_3
markupsafe=1.1.1=py37h8f50634_1
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py37hd478181_0
mistune=0.8.4=py37h8f50634_1001
mkl=2020.2=256
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
ml-collections=0.1.0=pypi_0
nbclient=0.5.0=py_0
nbconvert=6.0.7=py37hc8dfbb8_0
nbformat=5.0.7=py_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.4.1=py_0
ninja=1.10.1=py37hfd86e86_0
nodeenv=1.5.0=pypi_0
notebook=6.1.4=py37hc8dfbb8_0
numba=0.52.0=pypi_0
numpy=1.19.1=py37hbc911f0_0
numpy-base=1.19.1=py37hfa32c7d_0
nvidia-ml=7.352.0=py_0
olefile=0.46=py37_0
openssl=1.1.1i=h27cfd23_0
packaging=20.4=pyh9f0ad1d_0
pandas=1.2.0=pypi_0
pandoc=2.10.1=h516909a_0
pandocfilters=1.4.2=py_1
parso=0.7.1=pyh9f0ad1d_0
pathtools=0.1.2=py_1
patsy=0.5.1=pypi_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=py37hc8dfbb8_1
pickleshare=0.7.5=py37hc8dfbb8_1001
pillow=7.2.0=py37hb39fc2d_0
pip=20.2.3=py37_0
plotly=4.13.0=pypi_0
pre-commit=2.7.1=pypi_0
prometheus_client=0.8.0=pyh9f0ad1d_0
promise=2.3=py37hc8dfbb8_1
prompt-toolkit=3.0.7=py_0
prompt_toolkit=3.0.7=0
protobuf=3.13.0.1=py37h3340039_0
psutil=5.7.2=py37h8f50634_0
ptyprocess=0.6.0=py_1001
pycparser=2.20=pyh9f0ad1d_2
pygments=2.7.1=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.9.2=py37h05f1152_2
pyrsistent=0.17.3=py37h8f50634_0
pysocks=1.7.1=py37hc8dfbb8_1
python=3.7.9=h7579374_0
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.7.0=py3.7_cuda10.1.243_cudnn7.6.3_0
pytz=2020.5=pypi_0
pyyaml=5.3.1=py37h8f50634_0
pyzmq=19.0.2=py37hac76be4_1
qt=5.9.7=h5867ecd_1
qtconsole=4.7.7=py_0
qtpy=1.9.0=py_0
readline=8.0=h7b6447c_0
requests=2.24.0=pyh9f0ad1d_0
retrying=1.3.3=pypi_0
scikit-learn=0.23.2=py37h0573a6f_0
scipy=1.1.0=py37hfa4b5c9_1
send2trash=1.5.0=py_0
sentry-sdk=0.18.0=pyh9f0ad1d_0
setuptools=50.3.0=py37hb0f4dca_1
shortuuid=1.0.1=py37hc8dfbb8_2
sip=4.19.8=py37hf484d3e_0
six=1.15.0=py_0
sktime=0.5.1=pypi_0
smmap=3.0.4=pyh9f0ad1d_0
sqlite=3.33.0=h62c20be_0
statsmodels=0.12.1=pypi_0
subprocess32=3.5.4=pypi_0
terminado=0.9.1=py37hc8dfbb8_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
toml=0.10.1=pypi_0
torchaudio=0.7.0=py37
torchvision=0.8.1=py37_cu101
tornado=6.0.4=py37h8f50634_1
traitlets=5.0.4=py_1
typing_extensions=3.7.4.3=py_0
urllib3=1.25.10=py_0
virtualenv=20.0.33=pypi_0
wandb=0.10.11=pypi_0
watchdog=0.10.3=py37hc8dfbb8_1
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
wheel=0.35.1=py_0
widgetsnbextension=3.5.1=py37_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h516909a_0
zeromq=4.3.3=he1b5a44_2
zipp=3.3.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
